Partially working ddev_hostname command
