+++ cmd/ddev_hostname/ddev_hostname.go
@@ -1,8 +1,6 @@
 package main
 
 import (
-	"github.com/ddev/ddev/cmd/ddev/cmd"
-	"github.com/ddev/ddev/pkg/ddevapp"
 	"github.com/ddev/ddev/pkg/hostname"
 	"github.com/ddev/ddev/pkg/util"
 	"github.com/spf13/cobra"
@@ -14,17 +12,6 @@ import (
 func main() {
 	RootCmd.Execute()
 }
-
-// RootCmd represents the base command when called without any subcommands
-//var RootCmd = &cobra.Command{
-//	Use:     "ddev_hostname",
-//	Short:   "Manage hostnames in hosts file",
-//	Version: versionconstants.DdevVersion,
-//	PersistentPreRun: func(cmd *cobra.Command, args []string) {
-//		command := os.Args[1]
-//		_ = command // This is a placeholder to avoid unused variable error
-//	},
-//}
 
 // Execute adds all child commands to the root command sets flags appropriately.
 // This is called by main.main(). It only needs to happen once to the rootCmd.
@@ -43,14 +30,14 @@ var RootCmd = &cobra.Command{
 	Use:   "hostname [hostname] [ip]",
 	Short: "Manage your hostfile entries.",
 	Example: `
-ddev hostname junk.example.com 127.0.0.1
-ddev hostname -r junk.example.com 127.0.0.1
-ddev hostname --check junk.example.com 127.0.0.1
-ddev hostname --remove-inactive
+ddev_hostname hostname junk.example.com 127.0.0.1
+ddev_hostname hostname -r junk.example.com 127.0.0.1
+ddev_hostname hostname --check junk.example.com 127.0.0.1
+ddev_hostname hostname --remove-inactive
 `,
 	Long: `Manage your hostfile entries. Managing host names has security and usability
 implications and requires elevated privileges. You may be asked for a password
-to allow DDEV to modify your hosts file. If you are connected to the internet and using the domain ddev.site this is generally not necessary, because the hosts file never gets manipulated.`,
+to allow ddev_hostname to modify your hosts file. If you are connected to the internet and using the domain ddev.site this is generally not necessary, because the hosts file never gets manipulated.`,
 	Run: func(_ *cobra.Command, args []string) {
 
 		// Unless DDEV_NONINTERACTIVE is set (tests) then we need to be admin
@@ -58,20 +45,22 @@ to allow DDEV to modify your hosts file. If you are connected to the internet an
 			util.Failed("'ddev hostname %s' must be run with administrator privileges", strings.Join(args, " "))
 		}
 
+		// TODO: Reimplement this, figure out how to know what's inactive
+		// This may not be useful and not need to be implemented.
 		// If requested, remove all inactive host names and exit
-		if removeInactiveFlag {
-			if len(args) > 0 {
-				util.Failed("Invalid arguments supplied. 'ddev hostname --remove-all' accepts no arguments.")
-			}
-
-			util.Warning("Attempting to remove inactive custom hostnames for projects which are registered but not running")
-			removeInactiveHostnames()
-			return
-		}
+		//if removeInactiveFlag {
+		//	if len(args) > 0 {
+		//		util.Failed("Invalid arguments supplied. 'ddev hostname --remove-all' accepts no arguments.")
+		//	}
+		//
+		//	util.Warning("Attempting to remove inactive custom hostnames for projects which are registered but not running")
+		//	//removeInactiveHostnames()
+		//	return
+		//}
 
 		// If operating on one host name, two arguments are required
 		if len(args) != 2 {
-			util.Failed("Invalid arguments supplied. Please use 'ddev hostname [hostname] [ip]'")
+			util.Failed("Invalid arguments supplied. Please use 'ddev_hostname hostname [hostname] [ip]'")
 		}
 
 		name, dockerIP := args[0], args[1]
@@ -86,7 +75,7 @@ to allow DDEV to modify your hosts file. If you are connected to the internet an
 			return
 		}
 		if checkHostnameFlag {
-			exists, err := ddevapp.IsHostnameInHostsFile(name)
+			exists, err := hostname.IsHostnameInHostsFile(name)
 			if exists {
 				return
 			}
@@ -105,26 +94,22 @@ to allow DDEV to modify your hosts file. If you are connected to the internet an
 }
 
 // removeInactiveHostnames will remove all host names except those current in use by active projects.
-func removeInactiveHostnames() {
-	apps, err := ddevapp.GetInactiveProjects()
-	if err != nil {
-		util.Warning("Unable to run GetInactiveProjects: %v", err)
-		return
-	}
-	for _, app := range apps {
-		err := app.RemoveHostsEntriesIfNeeded()
-		if err != nil {
-			util.Warning("Unable to remove hosts entries for project '%s': %v", app.Name, err)
-		}
-	}
-}
+//func removeInactiveHostnames() {
+//	apps, err := ddevapp.GetInactiveProjects()
+//	if err != nil {
+//		util.Warning("Unable to run GetInactiveProjects: %v", err)
+//		return
+//	}
+//	for _, app := range apps {
+//		err := app.RemoveHostsEntriesIfNeeded()
+//		if err != nil {
+//			util.Warning("Unable to remove hosts entries for project '%s': %v", app.Name, err)
+//		}
+//	}
+//}
 
 func init() {
 	RootCmd.Flags().BoolVarP(&removeHostnameFlag, "remove", "r", false, "Remove the provided host name - ip correlation")
 	RootCmd.Flags().BoolVarP(&checkHostnameFlag, "check", "c", false, "Check to see if provided hostname is already in hosts file")
-	RootCmd.Flags().BoolVarP(&removeInactiveFlag, "remove-inactive", "R", false, "Remove host names of inactive projects")
-	RootCmd.Flags().BoolVar(&removeInactiveFlag, "fire-bazooka", false, "Alias of --remove-inactive")
-	_ = RootCmd.Flags().MarkHidden("fire-bazooka")
-
-	cmd.RootCmd.AddCommand(RootCmd)
+	//RootCmd.Flags().BoolVarP(&removeInactiveFlag, "remove-inactive", "R", false, "Remove host names of inactive projects")
 }
